AttackCooldown
AttackCooldown fully implemented

+++ Ice-wall-2/Assets/Scripts/UnitData.cs
@@ -5,7 +5,7 @@ public class UnitData : ScriptableObject
 {
     public int Price;
 	public GameObject UnitPrefab;
-	public int Strength;
+	public float Strength;
 	public float AttackCooldown;
 	public int Speed;
 }
